fix: revise bonus iteration instructions
Update bonus iteration 3.2 instructions to align it with the preceding content by removing mention of _refactoring_ and _polymorphism_. The iteration in the current state doesn't allow for efficient practice of refactoring. As well, polymorphism is an OOP concept applied to classes and methods, and therefore it is not suitable to have it mentioned here.